:truck: :recycle: :sparkles: :checkered_flag: Renamed util module to data. Also general cleanup, mostly around Note class
