+++ src/index.js
@@ -3,4 +3,4 @@ export * from './elements'
 export * from './note'
 export * from './sections'
 export * from './validate'
-export * from './util'
+export * from './data'
